Add some logs, skip stuff if already configured

+++ configure.sh
@@ -1,9 +1,22 @@
 #!/usr/bin/env sh
 
-# Configure Neovim
 mkdir -p ~/.config
-ln -s $(pwd)/nvim ~/.config/nvim
+
+# Configure Neovim
+if [ -d ~/.config/nvim ]; then
+  echo "[SKIP] Neovim"
+elif [ -f ~/.config/nvim ]; then
+  echo "[WARN] A file exists in the Neovim config path, please check ~/.config/nvim"
+else
+  ln -s $(pwd)/nvim ~/.config/nvim
+  echo "[OK] Neovim"
+fi
 
 # Configure tmux
-cp .tmux.conf ~/.tmux.conf
+if [ -f ~/.tmux.conf ]; then
+  echo "[SKIP] tmux"
+else
+  cp .tmux.conf ~/.tmux.conf
+  echo "[OK] tmux"
+fi
